doc-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-doc-engine"
version = "0.1.0"
description = "Reusable AI-assisted Chinese long-document engine for DOCX/PDF delivery."
requires-python = ">=3.11"
dependencies = [
  "python-docx>=1.1.2",
  "PyYAML>=6.0.1",
]

[project.scripts]
build-doc = "doc_engine.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
